[libc][math] Refactor log10, log1p, log2 implementation to header-only in src/__support/math folder.
Part of #147386

in preparation for:
https://discourse.llvm.org/t/rfc-make-clang-builtin-math-functions-constexpr-with-llvm-libc-to-support-c-23-constexpr-math-functions/86450
