+++ libc/shared/math.h
@@ -72,6 +72,9 @@
 #include "math/ldexpf16.h"
 #include "math/llogbf.h"
 #include "math/log.h"
+#include "math/log10.h"
+#include "math/log1p.h"
+#include "math/log2.h"
 #include "math/logbf.h"
 #include "math/logbf128.h"
 #include "math/logbf16.h"
