+++ libc/test/shared/CMakeLists.txt
@@ -63,11 +63,14 @@ add_fp_unittest(
     libc.src.__support.math.hypotf
     libc.src.__support.math.ilogbf16
     libc.src.__support.math.ilogbl
+    libc.src.__support.math.ldexpf
     libc.src.__support.math.log
+    libc.src.__support.math.log10
+    libc.src.__support.math.log1p
+    libc.src.__support.math.log2
     libc.src.__support.math.logbf
     libc.src.__support.math.logbf128
     libc.src.__support.math.logbf16
-    libc.src.__support.math.ldexpf
     libc.src.__support.math.ldexpf128
     libc.src.__support.math.ldexpf16
     libc.src.__support.math.llogbf
